# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My Career Open API Overview",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "MyCareer Backend URL"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDetailDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EyeSpottedError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDetailDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "RegisterDetailDTO": {
        "required": [
          "email",
          "password",
          "repeatPassword",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 2147483647,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "maxLength": 2147483647,
            "minLength": 3,
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 1,
            "type": "string"
          },
          "repeatPassword": {
            "maxLength": 2147483647,
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "EyeSpottedError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "LoginDetailDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}