# src/main/java/com/faros/EyeSpotted/model/DTO/LoginDetailDTO.java
package com.faros.EyeSpotted.model.DTO;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class LoginDetailDTO {
    private String username;
    private String password;
}